make vm reactive
