+++ app/src/main/java/dev/leonlatsch/photok/telemetry/ui/TelemetryExplanationViewModel.kt
@@ -1,0 +1,44 @@
+/*
+ *   Copyright 2020-2026 Leon Latsch
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package dev.leonlatsch.photok.telemetry.ui
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import dagger.hilt.android.lifecycle.HiltViewModel
+import dev.leonlatsch.photok.settings.data.Config
+import dev.leonlatsch.photok.telemetry.domain.TelemetryEnabledByDefault
+import dev.leonlatsch.photok.telemetry.domain.TelemetryService
+import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.stateIn
+import javax.inject.Inject
+
+@HiltViewModel
+class TelemetryExplanationViewModel @Inject constructor(
+    private val config: Config,
+    private val telemetryService: TelemetryService,
+) : ViewModel() {
+
+    val enabled = config.valuesFlow.map {
+        it[Config.TELEMETRY_ENABLED] as? Boolean ?: TelemetryEnabledByDefault
+    }.stateIn(viewModelScope, SharingStarted.WhileSubscribed(), TelemetryEnabledByDefault)
+
+    fun updateTelemetryEnabled(enabled: Boolean) {
+        config.telemetryEnabled = enabled
+        telemetryService.setup()
+    }
+}
